Cleaned up unused version check logic
