+++ redash/tasks/__init__.py
@@ -1,6 +1,5 @@
 from .general import (
     record_event,
-    version_check,
     send_mail,
     sync_user_details,
 )
@@ -19,4 +18,3 @@ from rq.connections import push_connection, pop_connection
 def init_app(app):
     app.before_request(lambda: push_connection(rq_redis_connection))
     app.teardown_request(lambda _: pop_connection())
-
