+++ redash/tasks/general.py
@@ -2,7 +2,6 @@ import requests
 from flask_mail import Message
 from redash import mail, models, settings
 from redash.models import users
-from redash.version_check import run_version_check
 from redash.worker import job, get_job_logger
 from redash.tasks.worker import Queue
 from redash.query_runner import NotSupported
@@ -27,27 +26,6 @@ def record_event(raw_event):
                 logger.error("Failed posting to %s: %s", hook, response.content)
         except Exception:
             logger.exception("Failed posting to %s", hook)
-
-
-def version_check():
-    run_version_check()
-
-
-@job("default")
-def subscribe(form):
-    logger.info(
-        "Subscribing to: [security notifications=%s], [newsletter=%s]",
-        form["security_notifications"],
-        form["newsletter"],
-    )
-    data = {
-        "admin_name": form["name"],
-        "admin_email": form["email"],
-        "org_name": form["org_name"],
-        "security_notifications": form["security_notifications"],
-        "newsletter": form["newsletter"],
-    }
-    requests.post("https://beacon.redash.io/subscribe", json=data)
 
 
 @job("emails")
